# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/visitors": {
      "post": {
        "operationId": "VisitorsController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVisitorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Visitor created successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create a new visitor",
        "tags": [
          "Visitors"
        ]
      },
      "get": {
        "operationId": "VisitorsController_findAll_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term across searchable fields",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Filter by creation start date (created_at >= startDate)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Filter by creation end date (created_at <= endDate)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entryGateId",
            "required": false,
            "in": "query",
            "description": "Filter by entry gate ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exitGateId",
            "required": false,
            "in": "query",
            "description": "Filter by exit gate ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return paginated visitors"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all visitors with pagination, search, sorting, and filters",
        "tags": [
          "Visitors"
        ]
      }
    },
    "/api/v1/visitors/export": {
      "get": {
        "operationId": "VisitorsController_exportToExcel_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term across searchable fields",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Filter by creation start date (created_at >= startDate)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Filter by creation end date (created_at <= endDate)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entryGateId",
            "required": false,
            "in": "query",
            "description": "Filter by entry gate ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exitGateId",
            "required": false,
            "in": "query",
            "description": "Filter by exit gate ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return exported visitors as Excel"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Export visitors to Excel with pagination, search, sorting, and filters",
        "tags": [
          "Visitors"
        ]
      }
    },
    "/api/v1/visitors/active": {
      "get": {
        "operationId": "VisitorsController_findAllActive_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term across searchable fields",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return paginated active visitors"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all active visitors inside the port",
        "tags": [
          "Visitors"
        ]
      }
    },
    "/api/v1/visitors/{id}": {
      "get": {
        "operationId": "VisitorsController_findOne_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return visitor"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Visitor not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get visitor by ID",
        "tags": [
          "Visitors"
        ]
      },
      "patch": {
        "operationId": "VisitorsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVisitorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Visitor updated successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Visitor not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update visitor",
        "tags": [
          "Visitors"
        ]
      },
      "delete": {
        "operationId": "VisitorsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Visitor deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Visitor not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete visitor (Super Admin only)",
        "tags": [
          "Visitors"
        ]
      }
    },
    "/api/v1/visitors/entry": {
      "post": {
        "operationId": "VisitorsController_registerEntry_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterVisitorEntryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Visitor entry registered successfully"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Register a visitor entry",
        "tags": [
          "Visitors"
        ]
      }
    },
    "/api/v1/visitors/{id}/exit": {
      "post": {
        "operationId": "VisitorsController_registerExit_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterVisitorExitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Visitor exit registered successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Visitor not found or not currently entered"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Register a visitor exit",
        "tags": [
          "Visitors"
        ]
      }
    },
    "/api/v1/visitors/{id}/cancel": {
      "post": {
        "operationId": "VisitorsController_cancel_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Visitor cancelled successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Visitor not found or already exited"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Cancel a visitor record",
        "tags": [
          "Visitors"
        ]
      }
    },
...
  },
  "info": {
    "title": "Terminal Port Management System (TPMS) API",
    "description": "API for managing and monitoring vehicles and visitors entering/exiting a terminal port, with JWT authentication and RBAC",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateVisitorDto": {
        "type": "object",
        "properties": {
          "visitorName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "nrcOrPassport": {
            "type": "string",
            "example": "12/ABC(N)123456"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+959123456789"
          },
          "companyName": {
            "type": "string",
            "example": "ABC Company"
          },
          "purposeOfVisit": {
            "type": "string",
            "example": "Meeting"
          },
          "hostEmployee": {
            "type": "string",
            "example": "Mr. Manager"
          },
          "entryGateId": {
            "type": "string"
          },
          "remarks": {
            "type": "string"
          }
        },
        "required": [
          "visitorName"
        ]
      },
      "RegisterVisitorEntryDto": {
        "type": "object",
        "properties": {
          "visitorName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "nrcOrPassport": {
            "type": "string",
            "example": "12/ABC(N)123456"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+959123456789"
          },
          "companyName": {
            "type": "string",
            "example": "ABC Company"
          },
          "purposeOfVisit": {
            "type": "string",
            "example": "Meeting"
          },
          "hostEmployee": {
            "type": "string",
            "example": "Mr. Manager"
          },
          "entryGateId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "remarks": {
            "type": "string"
          }
        },
        "required": [
          "visitorName",
          "entryGateId"
        ]
      },
      "RegisterVisitorExitDto": {
        "type": "object",
        "properties": {
          "exitGateId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "remarks": {
            "type": "string"
          }
        },
        "required": [
          "exitGateId"
        ]
      },
      "UpdateVisitorDto": {
        "type": "object",
        "properties": {
          "visitorName": {
            "type": "string",
            "example": "Jane Doe"
          },
          "nrcOrPassport": {
            "type": "string",
            "example": "12/ABC(N)123456"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+959123456789"
          },
          "companyName": {
            "type": "string",
            "example": "ABC Company"
          },
          "purposeOfVisit": {
            "type": "string",
            "example": "Meeting"
          },
          "hostEmployee": {
            "type": "string",
            "example": "Mr. Manager"
          },
          "entryGateId": {
            "type": "string"
          },
          "remarks": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ENTERED",
              "EXITED",
              "CANCELLED"
            ]
          },
          "exitGateId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}